Add calculator page
