+++ next.config.js
@@ -1,11 +1,7 @@
 module.exports = {
   swcMinify: false,
-  trailingSlash: true,
+  trailingSlash: false,
   env: {
-    // HOST
-    HOST_API_KEY: 'https://api-dev-minimal-v4.vercel.app',
-    // MAPBOX
-    MAPBOX_API: '',
     // FIREBASE
     FIREBASE_API_KEY: process.env.API_KEY,
     FIREBASE_AUTH_DOMAIN: process.env.AUTH_DOMAIN,
@@ -13,12 +9,5 @@ module.exports = {
     FIREBASE_STORAGE_BUCKET: process.env.STORAGE_BUCKET,
     FIREBASE_MESSAGING_SENDER_ID: process.env.MESSAGING_SENDER_ID,
     FIREBASE_APPID: process.env.APPID,
-    FIREBASE_MEASUREMENT_ID: process.env.MEASUREMENT_ID,
-    // AWS COGNITO
-    AWS_COGNITO_USER_POOL_ID: '',
-    AWS_COGNITO_CLIENT_ID: '',
-    // AUTH0
-    AUTH0_DOMAIN: '',
-    AUTH0_CLIENT_ID: '',
   },
 };
