+++ src/components/shared/settings/config-setting.ts
@@ -1,14 +1,10 @@
-// components
 import { SettingsValueProps } from './types';
-
-// PLEASE REMOVE `LOCAL STORAGE` WHEN YOU CHANGE SETTINGS.
-// ----------------------------------------------------------------------
 
 export const defaultSettings: SettingsValueProps = {
   themeMode: 'light',
   themeDirection: 'ltr',
   themeContrast: 'default',
   themeLayout: 'vertical',
-  themeColorPresets: 'default',
+  themeColorPresets: 'blue',
   themeStretch: false,
 };
