+++ src/hooks/useActiveLink.ts
@@ -12,16 +12,17 @@ export default function useActiveLink(path: string, deep = true): ReturnType {
 
   const checkPath = path.startsWith('#');
 
-  const currentPath = path === '/' ? '/' : `${path}/`;
+  const currentPath = path === '/' ? '/' : `${path}`;
 
   const normalActive =
     (!checkPath && pathname === currentPath) || (!checkPath && asPath === currentPath);
 
   const deepActive =
-    (!checkPath && pathname.includes(currentPath)) || (!checkPath && asPath.includes(currentPath));
+    (!checkPath && pathname.startsWith(currentPath)) ||
+    (!checkPath && asPath.startsWith(currentPath));
 
   return {
     active: deep ? deepActive : normalActive,
-    isExternalLink: path.includes('http'),
+    isExternalLink: path.startsWith('http'),
   };
 }
