+++ src/routes/paths.ts
@@ -20,9 +20,9 @@ export const PATH_DASHBOARD = {
   root: ROOTS_DASHBOARD,
   home: path(ROOTS_DASHBOARD, '/home'),
   calculator: {
-    root: path(ROOTS_DASHBOARD, '/calculator'),
-    list: path(ROOTS_DASHBOARD, '/calculator/list'),
-    new: path(ROOTS_DASHBOARD, '/calculator/new'),
-    view: (id: string) => path(ROOTS_DASHBOARD, `/calculator/${id}`),
+    root: path(ROOTS_DASHBOARD, '/import-calculator'),
+    list: path(ROOTS_DASHBOARD, '/import-calculator/list'),
+    new: path(ROOTS_DASHBOARD, '/import-calculator/new'),
+    view: (id: string) => path(ROOTS_DASHBOARD, `/import-calculator/${id}`),
   },
 };
